feat: Use Dressing.nvim for the LSP's code actions
Use Dressing.nvim as the backend of LSP code actions.
And, remove `lspsaga.nvim` as a result, as it'll no longer needed.
As a result, it's UI uses a Telescope-like picker thus, making it
prettier than the latter.
On the reason I've done it is `lspsaga.nvim` has a history of breaking!
changes, thus it would be safe to assume to replace it as a more
"stable" one.
`dressing.nvim` has a pretty UI andtakes advantages of `telescope.nvim`
and still being very performant.
As a side effect, we could not preview the changes which is unfortunate.
I'll go and try to fix it in a later commit.

Here is a list of the things changed:
- Removed `lspsaga.nvim`
- Defered the loading of `dressing.nvim` to `VeryLazy`.
- Use the default options of dressing.nvim
- Added a keymap for code actions.

And, here's a list of some problems:
- No code action preview

ADDITIONAL COMMIT:

feat: Add a description to the code_action keymap

I've added a description to the keymap which opens the buffer's current
code actions.
This will be useful if I were decide to add back `which-key.nvim`.

THINGS CHANGEDL
- Add a description to the code_action keymap

refactor: move dressing.nvim's plugin spec into the UI spec

Moving `dressing.nvim` from the `core` spec into the `ui` spec would fit
the plugins functionality, since `dressing.nvim` is a UI plugin.

THINGS CHANGED:
- Remove `dressing.nvim` from the `core` plugin spec
- Add `dressing.nvim` to the `ui` plugin spec