refactor: Move the specs.nvim config file inside the plugin spec
This is an unneeded require and actually might degrade readability thus,
I've removed the configuration file and moved it inside the plugin spec.

As for performance, some people might wonder if this is a bad thing for
performance because of the increased file size but, functions are
usually more expensive than arrays and, doing it this method is more
readable.