perf(gopher): remove
I really don't code in Go and actually the last time I was doing some
Go, I didn't have a decnet setup at all and I added this even without
testing.
So, of course that means this might not be the best solution thus, I'll
be removing this for the time being.