refactor!: Remove the inconsistent settings system
Before this commit, we had options to toggle plugins on/off.
But now, it has been a pain to add new plugins and, I don't like it.
Thus, I'm going to remove the whole system.
Anyone who wants to remove a plugin should toggle the `enabled` settings
provided by Lazy.
If anyone depended on this settings feature or like to see it back,
should open an issue - I'll add it back but more complete.
I - honestly don't want it in the first place but here we go.